Fixed watch bug in consul source.
watchplan.Run is a blocking call. Since it was being called inside of
the newWatcher method, this prevented the newWatcher method from
returning. Consul also immediately delivers a first change set to the
handler, which blocks indefinitely on sending the changeset, because
nothing can listen on the watcher channel until newWatcher completes.